refactor: 통합 확인후 toss, kakao config 삭제
